% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivar_norm_cp.R
\name{bivar_norm_cp}
\alias{bivar_norm_cp}
\title{Bivariate Normal Change Point Detection}
\usage{
bivar_norm_cp(
  x1,
  x2,
  theta = 0.95,
  lambda = 0.01,
  neu = 4,
  mu01 = 0.5,
  mu02 = 0.5,
  th_cp = 0.5,
  save_output = FALSE,
  file_name = "bivnormcp_output.xlsx"
)
}
\arguments{
\item{x1}{Numeric vector for the first series.}

\item{x2}{Numeric vector for the second series.}

\item{theta}{Prior probability for change point (default 0.95).}

\item{lambda}{Prior precision (default 0.01).}

\item{neu}{Degrees of freedom (default 4).}

\item{mu01}{Prior mean for x1 (default 0.5).}

\item{mu02}{Prior mean for x2 (default 0.5).}

\item{th_cp}{Threshold for detecting change point (default 0.5).}

\item{save_output}{Logical. If TRUE, saves the result to an Excel file (default is FALSE).}

\item{file_name}{Name of the Excel file to save the output (used only if save_output = TRUE).}
}
\value{
A list with original series and a matrix of posterior probabilities and detected change point locations.
}
\description{
Detects change points in bivariate normal data streams.
}
