% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_correction.R
\name{index_correction}
\alias{index_correction}
\title{Correct Index Jumps Within a Threshold}
\usage{
index_correction(index, tolerance = 2)
}
\arguments{
\item{index}{A numeric vector of change point indices.}

\item{tolerance}{Integer. The maximum allowed difference between consecutive indices to be considered the same (default is 2).}
}
\value{
A cleaned vector of unique indices.
}
\description{
Cleans up a sequence of indices by merging nearby ones based on a user-defined distance threshold.
}
